Remove redundant CanSee defns
Add a reason to CanSee
Move AGENT_ADDED to the very end, because handlers may need the coordinate of placed entity (eg. Combat -> CanSee -> requires coordinate)
Only advance with dt = 0 if IDle paused.  Not clear that it's desireable otherwise.. it's undesireable in the case of CONSOLE, whic was still advancing the time debt but failing to launch new verbs due to Verb:CanDo's check for pause.